more compiler warning fixes
git-svn-id: svn+ssh://svn.pointclouds.org/pcl/trunk@5066 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ registration/include/pcl/registration/impl/correspondence_estimation.hpp
@@ -90,7 +90,7 @@ pcl::registration::CorrespondenceEstimation<PointSource, PointTarget>::determine
     {
       if (distance[0] <= max_dist_sqr)
       {
-        corr.index_query = i;
+        corr.index_query = static_cast<int> (i);
         corr.index_match = index[0];
         corr.distance = distance[0];
         correspondences[i] = corr;
